api/requirements.txt
psycopg2
fastapi
python-jose[cryptography]
passlib[bcrypt]